Online News Popularity MLR Model
The reason why we chose this topic is because predicting the popularity of internet news has a wide range of applications nowadays. We want to get a deeper understanding of  the people who read online news. Consequently, it allows news organizations to deliver more relevant and compelling content and the company can allocate resources more wisely to prepare stories over their life cycle. Additionally, the prediction of news content is also beneficial for trend forecasting, and in this way can help advertisers propose more proﬁtable monetization techniques, and assist  readers ﬁlter the huge amount of information quickly and efﬁciently.